sea_query::expr

Trait ExprTrait

pub trait ExprTrait: Sized {

    // Required methods
    fn as_enum<N>(self, type_name: N) -> Expr
       where N: IntoIden;
    fn binary<O, R>(self, op: O, right: R) -> Expr
       where O: Into<BinOper>,
             R: Into<Expr>;
    fn cast_as<N>(self, type_name: N) -> Expr
       where N: IntoIden;
    fn unary(self, o: UnOper) -> Expr;
    fn max(self) -> Expr;
    fn min(self) -> Expr;
    fn sum(self) -> Expr;
    fn avg(self) -> Expr;
    fn count(self) -> Expr;
    fn count_distinct(self) -> Expr;
    fn if_null<V>(self, v: V) -> Expr
       where V: Into<Expr>;

    // Provided methods
    fn add<R>(self, right: R) -> Expr
       where R: Into<Expr> { ... }
    fn and<R>(self, right: R) -> Expr
       where R: Into<Expr> { ... }
    fn between<A, B>(self, a: A, b: B) -> Expr
       where A: Into<Expr>,
             B: Into<Expr> { ... }
    fn div<R>(self, right: R) -> Expr
       where R: Into<Expr> { ... }
    fn eq<R>(self, right: R) -> Expr
       where R: Into<Expr> { ... }
    fn equals<C>(self, col: C) -> Expr
       where C: IntoColumnRef { ... }
    fn gt<R>(self, right: R) -> Expr
       where R: Into<Expr> { ... }
    fn gte<R>(self, right: R) -> Expr
       where R: Into<Expr> { ... }
    fn in_subquery(self, sel: SelectStatement) -> Expr { ... }
    fn in_tuples<V, I>(self, v: I) -> Expr
       where V: IntoValueTuple,
             I: IntoIterator<Item = V> { ... }
    fn is<R>(self, right: R) -> Expr
       where R: Into<Expr> { ... }
    fn is_in<V, I>(self, v: I) -> Expr
       where V: Into<Expr>,
             I: IntoIterator<Item = V> { ... }
    fn is_not<R>(self, right: R) -> Expr
       where R: Into<Expr> { ... }
    fn is_not_in<V, I>(self, v: I) -> Expr
       where V: Into<Expr>,
             I: IntoIterator<Item = V> { ... }
    fn is_not_null(self) -> Expr { ... }
    fn is_null(self) -> Expr { ... }
    fn left_shift<R>(self, right: R) -> Expr
       where R: Into<Expr> { ... }
    fn like<L>(self, like: L) -> Expr
       where L: IntoLikeExpr { ... }
    fn lt<R>(self, right: R) -> Expr
       where R: Into<Expr> { ... }
    fn lte<R>(self, right: R) -> Expr
       where R: Into<Expr> { ... }
    fn modulo<R>(self, right: R) -> Expr
       where R: Into<Expr> { ... }
    fn mul<R>(self, right: R) -> Expr
       where R: Into<Expr> { ... }
    fn ne<R>(self, right: R) -> Expr
       where R: Into<Expr> { ... }
    fn not(self) -> Expr { ... }
    fn not_between<A, B>(self, a: A, b: B) -> Expr
       where A: Into<Expr>,
             B: Into<Expr> { ... }
    fn not_equals<C>(self, col: C) -> Expr
       where C: IntoColumnRef { ... }
    fn not_in_subquery(self, sel: SelectStatement) -> Expr { ... }
    fn not_like<L>(self, like: L) -> Expr
       where L: IntoLikeExpr { ... }
    fn or<R>(self, right: R) -> Expr
       where R: Into<Expr> { ... }
    fn right_shift<R>(self, right: R) -> Expr
       where R: Into<Expr> { ... }
    fn sub<R>(self, right: R) -> Expr
       where R: Into<Expr> { ... }
    fn bit_and<R>(self, right: R) -> Expr
       where R: Into<Expr> { ... }
    fn bit_or<R>(self, right: R) -> Expr
       where R: Into<Expr> { ... }

}

“Operator” methods for building expressions.

Before sea_query 0.32.0 (sea_orm 1.1.1), these methods were awailable only on Expr/SimpleExpr and you needed to manually construct these types first.

Now, you can call them directly on any expression type:

let expr = 1_i32.cast_as("REAL");
let expr = Func::char_length("abc").eq(3_i32);
let expr = Expr::current_date().cast_as("TEXT").like("2024%");

If some methods are missing, look into BinOper::Custom, Func::cust, or Expr::cust* as a workaround, and consider reporting your issue.

Required Methods

fn as_enum<N>(self, type_name: N) -> Expr
where N: IntoIden,

Express a AS enum expression.

Examples

use sea_query::{tests_cfg::*, *};

let query = Query::insert()
    .into_table(Char::Table)
    .columns([Char::FontSize])
    .values_panic(["large".as_enum("FontSizeEnum")])
    .to_owned();

assert_eq!(
    query.to_string(MsSqlQueryBuilder),
    r#"INSERT INTO [character] ([font_size]) VALUES ('large')"#
);
assert_eq!(
    query.to_string(MysqlQueryBuilder),
    r#"INSERT INTO `character` (`font_size`) VALUES ('large')"#
);
assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"INSERT INTO "character" ("font_size") VALUES (CAST('large' AS "FontSizeEnum"))"#
);
assert_eq!(
    query.to_string(SqliteQueryBuilder),
    r#"INSERT INTO "character" ("font_size") VALUES ('large')"#
);

let query = Query::select()
    .expr(Expr::col(Char::FontSize).as_enum("FontSizeEnum[]"))
    .from(Char::Table)
    .to_owned();

assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT CAST("font_size" AS "FontSizeEnum"[]) FROM "character""#
);

fn binary<O, R>(self, op: O, right: R) -> Expr
where O: Into<BinOper>, R: Into<Expr>,

Create any binary operation

Examples

use sea_query::{*, tests_cfg::*};

let query = Query::select()
    .columns([Char::Character, Char::SizeW, Char::SizeH])
    .from(Char::Table)
    .cond_where(all![
        Char::SizeW.into_column_ref().binary(BinOper::SmallerThan, 10),
        Char::SizeW.into_column_ref().binary(BinOper::GreaterThan, Char::SizeH.into_column_ref())
    ])
    .to_owned();
assert_eq!(
    query.to_string(MsSqlQueryBuilder),
    r#"SELECT [character], [size_w], [size_h] FROM [character] WHERE [size_w] < 10 AND [size_w] > [size_h]"#
);
assert_eq!(
    query.to_string(MysqlQueryBuilder),
    r#"SELECT `character`, `size_w`, `size_h` FROM `character` WHERE `size_w` < 10 AND `size_w` > `size_h`"#
);
assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE "size_w" < 10 AND "size_w" > "size_h""#
);
assert_eq!(
    query.to_string(SqliteQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE "size_w" < 10 AND "size_w" > "size_h""#
);

fn cast_as<N>(self, type_name: N) -> Expr
where N: IntoIden,

Express a CAST AS expression.

Examples

use sea_query::{tests_cfg::*, *};

let query = Query::select().expr("1".cast_as("integer")).to_owned();

assert_eq!(
    query.to_string(MsSqlQueryBuilder),
    r#"SELECT CAST('1' AS integer)"#
);
assert_eq!(
    query.to_string(MysqlQueryBuilder),
    r#"SELECT CAST('1' AS integer)"#
);
assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT CAST('1' AS integer)"#
);
assert_eq!(
    query.to_string(SqliteQueryBuilder),
    r#"SELECT CAST('1' AS integer)"#
);

fn unary(self, o: UnOper) -> Expr

Apply any unary operator to the expression.

Examples

use sea_query::{*, tests_cfg::*};

let query = Query::select()
    .columns([Char::Character, Char::SizeW, Char::SizeH])
    .from(Char::Table)
    .and_where(Expr::col((Char::Table, Char::SizeW)).is_null().unary(UnOper::Not))
    .to_owned();

assert_eq!(
    query.to_string(MsSqlQueryBuilder),
    r#"SELECT [character], [size_w], [size_h] FROM [character] WHERE NOT [character].[size_w] IS NULL"#
);
assert_eq!(
    query.to_string(MysqlQueryBuilder),
    r#"SELECT `character`, `size_w`, `size_h` FROM `character` WHERE NOT `character`.`size_w` IS NULL"#
);
assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE NOT "character"."size_w" IS NULL"#
);
assert_eq!(
    query.to_string(SqliteQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE NOT "character"."size_w" IS NULL"#
);

fn max(self) -> Expr

Express a MAX function.

Examples

use sea_query::{tests_cfg::*, *};

let query = Query::select()
    .expr(Expr::col((Char::Table, Char::SizeW)).max())
    .from(Char::Table)
    .to_owned();

assert_eq!(
    query.to_string(MsSqlQueryBuilder),
    r#"SELECT MAX([character].[size_w]) FROM [character]"#
);
assert_eq!(
    query.to_string(MysqlQueryBuilder),
    r#"SELECT MAX(`character`.`size_w`) FROM `character`"#
);
assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT MAX("character"."size_w") FROM "character""#
);
assert_eq!(
    query.to_string(SqliteQueryBuilder),
    r#"SELECT MAX("character"."size_w") FROM "character""#
);

fn min(self) -> Expr

Express a MIN function.

Examples

use sea_query::{tests_cfg::*, *};

let query = Query::select()
    .expr(Expr::col((Char::Table, Char::SizeW)).min())
    .from(Char::Table)
    .to_owned();

assert_eq!(
    query.to_string(MsSqlQueryBuilder),
    r#"SELECT MIN([character].[size_w]) FROM [character]"#
);
assert_eq!(
    query.to_string(MysqlQueryBuilder),
    r#"SELECT MIN(`character`.`size_w`) FROM `character`"#
);
assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT MIN("character"."size_w") FROM "character""#
);
assert_eq!(
    query.to_string(SqliteQueryBuilder),
    r#"SELECT MIN("character"."size_w") FROM "character""#
);

fn sum(self) -> Expr

Express a SUM function.

Examples

use sea_query::{tests_cfg::*, *};

let query = Query::select()
    .expr(Expr::col((Char::Table, Char::SizeW)).sum())
    .from(Char::Table)
    .to_owned();

assert_eq!(
    query.to_string(MsSqlQueryBuilder),
    r#"SELECT SUM([character].[size_w]) FROM [character]"#
);
assert_eq!(
    query.to_string(MysqlQueryBuilder),
    r#"SELECT SUM(`character`.`size_w`) FROM `character`"#
);
assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT SUM("character"."size_w") FROM "character""#
);
assert_eq!(
    query.to_string(SqliteQueryBuilder),
    r#"SELECT SUM("character"."size_w") FROM "character""#
);

fn avg(self) -> Expr

Express a AVG (average) function.

Examples

use sea_query::{tests_cfg::*, *};

let query = Query::select()
    .expr(Expr::col((Char::Table, Char::SizeW)).avg())
    .from(Char::Table)
    .to_owned();

assert_eq!(
    query.to_string(MsSqlQueryBuilder),
    r#"SELECT AVG([character].[size_w]) FROM [character]"#
);
assert_eq!(
    query.to_string(MysqlQueryBuilder),
    r#"SELECT AVG(`character`.`size_w`) FROM `character`"#
);
assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT AVG("character"."size_w") FROM "character""#
);
assert_eq!(
    query.to_string(SqliteQueryBuilder),
    r#"SELECT AVG("character"."size_w") FROM "character""#
);

fn count(self) -> Expr

Express a COUNT function.

Examples

use sea_query::{tests_cfg::*, *};

let query = Query::select()
    .expr(Expr::col((Char::Table, Char::SizeW)).count())
    .from(Char::Table)
    .to_owned();

assert_eq!(
    query.to_string(MsSqlQueryBuilder),
    r#"SELECT COUNT([character].[size_w]) FROM [character]"#
);
assert_eq!(
    query.to_string(MysqlQueryBuilder),
    r#"SELECT COUNT(`character`.`size_w`) FROM `character`"#
);
assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT COUNT("character"."size_w") FROM "character""#
);
assert_eq!(
    query.to_string(SqliteQueryBuilder),
    r#"SELECT COUNT("character"."size_w") FROM "character""#
);

fn count_distinct(self) -> Expr

Express a COUNT function with the DISTINCT modifier.

Examples

use sea_query::{tests_cfg::*, *};

let query = Query::select()
    .expr(Expr::col((Char::Table, Char::SizeW)).count_distinct())
    .from(Char::Table)
    .to_owned();

assert_eq!(
    query.to_string(MsSqlQueryBuilder),
    r#"SELECT COUNT(DISTINCT [character].[size_w]) FROM [character]"#
);
assert_eq!(
    query.to_string(MysqlQueryBuilder),
    r#"SELECT COUNT(DISTINCT `character`.`size_w`) FROM `character`"#
);
assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT COUNT(DISTINCT "character"."size_w") FROM "character""#
);
assert_eq!(
    query.to_string(SqliteQueryBuilder),
    r#"SELECT COUNT(DISTINCT "character"."size_w") FROM "character""#
);

fn if_null<V>(self, v: V) -> Expr
where V: Into<Expr>,

Express a IF NULL function.

Examples

use sea_query::{tests_cfg::*, *};

let query = Query::select()
    .expr(Expr::col((Char::Table, Char::SizeW)).if_null(0))
    .from(Char::Table)
    .to_owned();

assert_eq!(
    query.to_string(MsSqlQueryBuilder),
    r#"SELECT ISNULL([character].[size_w], 0) FROM [character]"#
);
assert_eq!(
    query.to_string(MysqlQueryBuilder),
    r#"SELECT IFNULL(`character`.`size_w`, 0) FROM `character`"#
);
assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT COALESCE("character"."size_w", 0) FROM "character""#
);
assert_eq!(
    query.to_string(SqliteQueryBuilder),
    r#"SELECT IFNULL("character"."size_w", 0) FROM "character""#
);

Provided Methods

fn add<R>(self, right: R) -> Expr
where R: Into<Expr>,

Express an arithmetic addition operation.

Examples

Adding literal values

use sea_query::{tests_cfg::*, *};

let query = Query::select()
    .columns([Char::Character, Char::SizeW, Char::SizeH])
    .from(Char::Table)
    .and_where(1.add(1).eq(2))
    .to_owned();

assert_eq!(
    query.to_string(MsSqlQueryBuilder),
    r#"SELECT [character], [size_w], [size_h] FROM [character] WHERE 1 + 1 = 2"#
);
assert_eq!(
    query.to_string(MysqlQueryBuilder),
    r#"SELECT `character`, `size_w`, `size_h` FROM `character` WHERE 1 + 1 = 2"#
);
assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE 1 + 1 = 2"#
);
assert_eq!(
    query.to_string(SqliteQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE 1 + 1 = 2"#
);

Adding columns and values

use sea_query::{tests_cfg::*, *};

let query = Query::select()
    .expr(Expr::col(Char::SizeW).add(1))
    .from(Char::Table)
    .to_owned();

assert_eq!(
    query.to_string(MysqlQueryBuilder),
    r#"SELECT `size_w` + 1 FROM `character`"#
);
assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT "size_w" + 1 FROM "character""#
);
assert_eq!(
    query.to_string(SqliteQueryBuilder),
    r#"SELECT "size_w" + 1 FROM "character""#
);

Adding columns

use sea_query::{tests_cfg::*, *};

let query = Query::select()
    .expr(Expr::col(Char::SizeW).add(Expr::col(Char::SizeH)))
    .from(Char::Table)
    .to_owned();

assert_eq!(
    query.to_string(MysqlQueryBuilder),
    r#"SELECT `size_w` + `size_h` FROM `character`"#
);
assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT "size_w" + "size_h" FROM "character""#
);
assert_eq!(
    query.to_string(SqliteQueryBuilder),
    r#"SELECT "size_w" + "size_h" FROM "character""#
);

fn and<R>(self, right: R) -> Expr
where R: Into<Expr>,

Express a logical AND operation.

Examples

use sea_query::{*, tests_cfg::*};

let query = Query::select()
    .columns([Char::Character, Char::SizeW, Char::SizeH])
    .from(Char::Table)
    .cond_where(any![
        Expr::col(Char::SizeW).eq(1).and(Expr::col(Char::SizeH).eq(2)),
        Expr::col(Char::SizeW).eq(3).and(Expr::col(Char::SizeH).eq(4)),
    ])
    .to_owned();

assert_eq!(
    query.to_string(MsSqlQueryBuilder),
    r#"SELECT [character], [size_w], [size_h] FROM [character] WHERE ([size_w] = 1 AND [size_h] = 2) OR ([size_w] = 3 AND [size_h] = 4)"#
);
assert_eq!(
    query.to_string(MysqlQueryBuilder),
    r#"SELECT `character`, `size_w`, `size_h` FROM `character` WHERE (`size_w` = 1 AND `size_h` = 2) OR (`size_w` = 3 AND `size_h` = 4)"#
);
assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE ("size_w" = 1 AND "size_h" = 2) OR ("size_w" = 3 AND "size_h" = 4)"#
);
assert_eq!(
    query.to_string(SqliteQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE ("size_w" = 1 AND "size_h" = 2) OR ("size_w" = 3 AND "size_h" = 4)"#
);

fn between<A, B>(self, a: A, b: B) -> Expr
where A: Into<Expr>, B: Into<Expr>,

Express a BETWEEN expression.

Examples

use sea_query::{*, tests_cfg::*};

let query = Query::select()
    .columns([Char::Character, Char::SizeW, Char::SizeH])
    .from(Char::Table)
    .and_where((Char::Table, Char::SizeW).into_column_ref().between(1, 10))
    .to_owned();

assert_eq!(
    query.to_string(MysqlQueryBuilder),
    r#"SELECT `character`, `size_w`, `size_h` FROM `character` WHERE `character`.`size_w` BETWEEN 1 AND 10"#
);
assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE "character"."size_w" BETWEEN 1 AND 10"#
);
assert_eq!(
    query.to_string(SqliteQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE "character"."size_w" BETWEEN 1 AND 10"#
);

fn div<R>(self, right: R) -> Expr
where R: Into<Expr>,

Express an arithmetic division operation.

Examples

use sea_query::{tests_cfg::*, *};

let query = Query::select()
    .columns([Char::Character, Char::SizeW, Char::SizeH])
    .from(Char::Table)
    .and_where(1.div(1).eq(2))
    .to_owned();

assert_eq!(
    query.to_string(MsSqlQueryBuilder),
    r#"SELECT [character], [size_w], [size_h] FROM [character] WHERE 1 / 1 = 2"#
);
assert_eq!(
    query.to_string(MysqlQueryBuilder),
    r#"SELECT `character`, `size_w`, `size_h` FROM `character` WHERE 1 / 1 = 2"#
);
assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE 1 / 1 = 2"#
);
assert_eq!(
    query.to_string(SqliteQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE 1 / 1 = 2"#
);

fn eq<R>(self, right: R) -> Expr
where R: Into<Expr>,

Express an equal (=) expression.

Examples

use sea_query::{tests_cfg::*, *};

let query = Query::select()
    .columns([Char::Character, Char::SizeW, Char::SizeH])
    .from(Char::Table)
    .and_where(Expr::val("What!").eq("Nothing"))
    .and_where(Char::Id.into_column_ref().eq(1))
    .to_owned();

assert_eq!(
    query.to_string(MsSqlQueryBuilder),
    r#"SELECT [character], [size_w], [size_h] FROM [character] WHERE 'What!' = 'Nothing' AND [id] = 1"#
);
assert_eq!(
    query.to_string(MysqlQueryBuilder),
    r#"SELECT `character`, `size_w`, `size_h` FROM `character` WHERE 'What!' = 'Nothing' AND `id` = 1"#
);
assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE 'What!' = 'Nothing' AND "id" = 1"#
);
assert_eq!(
    query.to_string(SqliteQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE 'What!' = 'Nothing' AND "id" = 1"#
);

Note how you should express a string being a literal vs an identifier.

use sea_query::{tests_cfg::*, *};

let query = Query::select()
    .column(Char::Character)
    .from(Char::Table)
    .and_where(Expr::col("name").eq("Something"))
    .to_owned();

assert_eq!(
    query.to_string(MysqlQueryBuilder),
    r#"SELECT `character` FROM `character` WHERE `name` = 'Something'"#
);
assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT "character" FROM "character" WHERE "name" = 'Something'"#
);
assert_eq!(
    query.to_string(SqliteQueryBuilder),
    r#"SELECT "character" FROM "character" WHERE "name" = 'Something'"#
);

fn equals<C>(self, col: C) -> Expr
where C: IntoColumnRef,

Express a equal expression between two table columns, you will mainly use this to relate identical value between two table columns.

Examples

use sea_query::{*, tests_cfg::*};

let query = Query::select()
    .columns([Char::Character, Char::SizeW, Char::SizeH])
    .from(Char::Table)
    .and_where((Char::Table, Char::FontId).into_column_ref().equals((Font::Table, Font::Id)))
    .to_owned();

assert_eq!(
    query.to_string(MsSqlQueryBuilder),
    r#"SELECT [character], [size_w], [size_h] FROM [character] WHERE [character].[font_id] = [font].[id]"#
);
assert_eq!(
    query.to_string(MysqlQueryBuilder),
    r#"SELECT `character`, `size_w`, `size_h` FROM `character` WHERE `character`.`font_id` = `font`.`id`"#
);
assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE "character"."font_id" = "font"."id""#
);
assert_eq!(
    query.to_string(SqliteQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE "character"."font_id" = "font"."id""#
);

fn gt<R>(self, right: R) -> Expr
where R: Into<Expr>,

Express a greater than (>) expression.

Examples

use sea_query::{tests_cfg::*, *};

let query = Query::select()
    .columns([Char::Character, Char::SizeW, Char::SizeH])
    .from(Char::Table)
    .and_where((Char::Table, Char::SizeW).into_column_ref().gt(2))
    .to_owned();

assert_eq!(
    query.to_string(MsSqlQueryBuilder),
    r#"SELECT [character], [size_w], [size_h] FROM [character] WHERE [character].[size_w] > 2"#
);
assert_eq!(
    query.to_string(MysqlQueryBuilder),
    r#"SELECT `character`, `size_w`, `size_h` FROM `character` WHERE `character`.`size_w` > 2"#
);
assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE "character"."size_w" > 2"#
);
assert_eq!(
    query.to_string(SqliteQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE "character"."size_w" > 2"#
);

fn gte<R>(self, right: R) -> Expr
where R: Into<Expr>,

Express a greater than or equal (>=) expression.

Examples

use sea_query::{tests_cfg::*, *};

let query = Query::select()
    .columns([Char::Character, Char::SizeW, Char::SizeH])
    .from(Char::Table)
    .and_where((Char::Table, Char::SizeW).into_column_ref().gte(2))
    .to_owned();

assert_eq!(
    query.to_string(MsSqlQueryBuilder),
    r#"SELECT [character], [size_w], [size_h] FROM [character] WHERE [character].[size_w] >= 2"#
);
assert_eq!(
    query.to_string(MysqlQueryBuilder),
    r#"SELECT `character`, `size_w`, `size_h` FROM `character` WHERE `character`.`size_w` >= 2"#
);
assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE "character"."size_w" >= 2"#
);
assert_eq!(
    query.to_string(SqliteQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE "character"."size_w" >= 2"#
);

fn in_subquery(self, sel: SelectStatement) -> Expr

Express a IN sub-query expression.

Examples

use sea_query::{*, tests_cfg::*};

let query = Query::select()
    .columns([Char::Character, Char::SizeW, Char::SizeH])
    .from(Char::Table)
    .and_where(Char::SizeW.into_column_ref().in_subquery(
        Query::select()
            .expr(Expr::cust("3 + 2 * 2"))
            .take()
    ))
    .to_owned();

assert_eq!(
    query.to_string(MsSqlQueryBuilder),
    r#"SELECT [character], [size_w], [size_h] FROM [character] WHERE [size_w] IN (SELECT 3 + 2 * 2)"#
);
assert_eq!(
    query.to_string(MysqlQueryBuilder),
    r#"SELECT `character`, `size_w`, `size_h` FROM `character` WHERE `size_w` IN (SELECT 3 + 2 * 2)"#
);
assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE "size_w" IN (SELECT 3 + 2 * 2)"#
);
assert_eq!(
    query.to_string(SqliteQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE "size_w" IN (SELECT 3 + 2 * 2)"#
);

fn in_tuples<V, I>(self, v: I) -> Expr
where V: IntoValueTuple, I: IntoIterator<Item = V>,

Express a IN sub expression.

Examples

use sea_query::{*, tests_cfg::*};

let query = Query::select()
    .columns([Char::Character, Char::FontId])
    .from(Char::Table)
    .and_where(
        Expr::tuple([
            Expr::col(Char::Character).into(),
            Expr::col(Char::FontId).into(),
        ]).in_tuples([(1, String::from("1")), (2, String::from("2"))])
    )
    .to_owned();

assert_eq!(
    query.to_string(MysqlQueryBuilder),
    r#"SELECT `character`, `font_id` FROM `character` WHERE (`character`, `font_id`) IN ((1, '1'), (2, '2'))"#
);

assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT "character", "font_id" FROM "character" WHERE ("character", "font_id") IN ((1, '1'), (2, '2'))"#
);

assert_eq!(
    query.to_string(SqliteQueryBuilder),
    r#"SELECT "character", "font_id" FROM "character" WHERE ("character", "font_id") IN ((1, '1'), (2, '2'))"#
);

fn is<R>(self, right: R) -> Expr
where R: Into<Expr>,

Express a IS expression.

Examples

use sea_query::{*, tests_cfg::*};

let query = Query::select()
    .columns([Char::Character, Char::SizeW, Char::SizeH])
    .from(Char::Table)
    .and_where((Char::Table, Char::Ascii).into_column_ref().is(true))
    .to_owned();

assert_eq!(
    query.to_string(MsSqlQueryBuilder),
    r#"SELECT [character], [size_w], [size_h] FROM [character] WHERE [character].[ascii] IS 1=1"#
);
assert_eq!(
    query.to_string(MysqlQueryBuilder),
    r#"SELECT `character`, `size_w`, `size_h` FROM `character` WHERE `character`.`ascii` IS TRUE"#
);
assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE "character"."ascii" IS TRUE"#
);
assert_eq!(
    query.to_string(SqliteQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE "character"."ascii" IS TRUE"#
);

fn is_in<V, I>(self, v: I) -> Expr
where V: Into<Expr>, I: IntoIterator<Item = V>,

Express a IN expression.

Examples

use sea_query::{tests_cfg::*, *};

let query = Query::select()
    .columns([Char::Id])
    .from(Char::Table)
    .and_where(
        (Char::Table, Char::SizeW)
            .into_column_ref()
            .is_in([1, 2, 3]),
    )
    .to_owned();

assert_eq!(
    query.to_string(MsSqlQueryBuilder),
    r#"SELECT [id] FROM [character] WHERE [character].[size_w] IN (1, 2, 3)"#
);
assert_eq!(
    query.to_string(MysqlQueryBuilder),
    r#"SELECT `id` FROM `character` WHERE `character`.`size_w` IN (1, 2, 3)"#
);
assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT "id" FROM "character" WHERE "character"."size_w" IN (1, 2, 3)"#
);
assert_eq!(
    query.to_string(SqliteQueryBuilder),
    r#"SELECT "id" FROM "character" WHERE "character"."size_w" IN (1, 2, 3)"#
);

The same query can be constructed using the raw_query! macro.

use sea_query::Values;

let ids: Vec<i32> = vec![1, 2, 3];
let query = sea_query::raw_query!(
    SqliteQueryBuilder,
    r#"SELECT "id" FROM "character" WHERE "character"."size_w" IN ({..ids})"#
);

assert_eq!(
    query.sql,
    r#"SELECT "id" FROM "character" WHERE "character"."size_w" IN (?, ?, ?)"#
);
assert_eq!(query.values, Values(vec![1.into(), 2.into(), 3.into()]));

Empty value list is converted to an always false expression due to SQL syntax.

use sea_query::{tests_cfg::*, *};

let query = Query::select()
    .columns([Char::Id])
    .from(Char::Table)
    .and_where(
        (Char::Table, Char::SizeW)
            .into_column_ref()
            .is_in(Vec::<u8>::new()),
    )
    .to_owned();

assert_eq!(
    query.to_string(MsSqlQueryBuilder),
    r#"SELECT [id] FROM [character] WHERE 1 = 2"#
);
assert_eq!(
    query.to_string(MysqlQueryBuilder),
    r#"SELECT `id` FROM `character` WHERE 1 = 2"#
);
assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT "id" FROM "character" WHERE 1 = 2"#
);
assert_eq!(
    query.to_string(SqliteQueryBuilder),
    r#"SELECT "id" FROM "character" WHERE 1 = 2"#
);

fn is_not<R>(self, right: R) -> Expr
where R: Into<Expr>,

Express a IS NOT expression.

Examples

use sea_query::{*, tests_cfg::*};

let query = Query::select()
    .columns([Char::Character, Char::SizeW, Char::SizeH])
    .from(Char::Table)
    .and_where((Char::Table, Char::Ascii).into_column_ref().is_not(true))
    .to_owned();

assert_eq!(
    query.to_string(MsSqlQueryBuilder),
    r#"SELECT [character], [size_w], [size_h] FROM [character] WHERE [character].[ascii] IS NOT 1=1"#
);
assert_eq!(
    query.to_string(MysqlQueryBuilder),
    r#"SELECT `character`, `size_w`, `size_h` FROM `character` WHERE `character`.`ascii` IS NOT TRUE"#
);
assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE "character"."ascii" IS NOT TRUE"#
);
assert_eq!(
    query.to_string(SqliteQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE "character"."ascii" IS NOT TRUE"#
);

fn is_not_in<V, I>(self, v: I) -> Expr
where V: Into<Expr>, I: IntoIterator<Item = V>,

Express a NOT IN expression.

Examples

use sea_query::{tests_cfg::*, *};

let query = Query::select()
    .columns([Char::Id])
    .from(Char::Table)
    .and_where(
        (Char::Table, Char::SizeW)
            .into_column_ref()
            .is_not_in([1, 2, 3]),
    )
    .to_owned();

assert_eq!(
    query.to_string(MsSqlQueryBuilder),
    r#"SELECT [id] FROM [character] WHERE [character].[size_w] NOT IN (1, 2, 3)"#
);
assert_eq!(
    query.to_string(MysqlQueryBuilder),
    r#"SELECT `id` FROM `character` WHERE `character`.`size_w` NOT IN (1, 2, 3)"#
);
assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT "id" FROM "character" WHERE "character"."size_w" NOT IN (1, 2, 3)"#
);
assert_eq!(
    query.to_string(SqliteQueryBuilder),
    r#"SELECT "id" FROM "character" WHERE "character"."size_w" NOT IN (1, 2, 3)"#
);

Empty value list

use sea_query::{tests_cfg::*, *};

let query = Query::select()
    .columns([Char::Id])
    .from(Char::Table)
    .and_where(
        (Char::Table, Char::SizeW)
            .into_column_ref()
            .is_not_in(Vec::<u8>::new()),
    )
    .to_owned();

assert_eq!(
    query.to_string(MsSqlQueryBuilder),
    r#"SELECT [id] FROM [character] WHERE 1 = 1"#
);
assert_eq!(
    query.to_string(MysqlQueryBuilder),
    r#"SELECT `id` FROM `character` WHERE 1 = 1"#
);
assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT "id" FROM "character" WHERE 1 = 1"#
);
assert_eq!(
    query.to_string(SqliteQueryBuilder),
    r#"SELECT "id" FROM "character" WHERE 1 = 1"#
);

fn is_not_null(self) -> Expr

Express a IS NOT NULL expression.

Examples

use sea_query::{*, tests_cfg::*};

let query = Query::select()
    .columns([Char::Character, Char::SizeW, Char::SizeH])
    .from(Char::Table)
    .and_where((Char::Table, Char::SizeW).into_column_ref().is_not_null())
    .to_owned();

assert_eq!(
    query.to_string(MsSqlQueryBuilder),
    r#"SELECT [character], [size_w], [size_h] FROM [character] WHERE [character].[size_w] IS NOT NULL"#
);
assert_eq!(
    query.to_string(MysqlQueryBuilder),
    r#"SELECT `character`, `size_w`, `size_h` FROM `character` WHERE `character`.`size_w` IS NOT NULL"#
);
assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE "character"."size_w" IS NOT NULL"#
);
assert_eq!(
    query.to_string(SqliteQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE "character"."size_w" IS NOT NULL"#
);

fn is_null(self) -> Expr

Express a IS NULL expression.

Examples

use sea_query::{*, tests_cfg::*};

let query = Query::select()
    .columns([Char::Character, Char::SizeW, Char::SizeH])
    .from(Char::Table)
    .and_where((Char::Table, Char::SizeW).into_column_ref().is_null())
    .to_owned();

assert_eq!(
    query.to_string(MsSqlQueryBuilder),
    r#"SELECT [character], [size_w], [size_h] FROM [character] WHERE [character].[size_w] IS NULL"#
);
assert_eq!(
    query.to_string(MysqlQueryBuilder),
    r#"SELECT `character`, `size_w`, `size_h` FROM `character` WHERE `character`.`size_w` IS NULL"#
);
assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE "character"."size_w" IS NULL"#
);
assert_eq!(
    query.to_string(SqliteQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE "character"."size_w" IS NULL"#
);

fn left_shift<R>(self, right: R) -> Expr
where R: Into<Expr>,

Express a bitwise left shift.

Examples

use sea_query::{tests_cfg::*, *};

let query = Query::select()
    .columns([Char::Character, Char::SizeW, Char::SizeH])
    .from(Char::Table)
    .and_where(1.left_shift(1).eq(2))
    .to_owned();

assert_eq!(
    query.to_string(MsSqlQueryBuilder),
    r#"SELECT [character], [size_w], [size_h] FROM [character] WHERE 1 << 1 = 2"#
);
assert_eq!(
    query.to_string(MysqlQueryBuilder),
    r#"SELECT `character`, `size_w`, `size_h` FROM `character` WHERE 1 << 1 = 2"#
);
assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE 1 << 1 = 2"#
);
assert_eq!(
    query.to_string(SqliteQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE 1 << 1 = 2"#
);

fn like<L>(self, like: L) -> Expr
where L: IntoLikeExpr,

Express a LIKE expression.

Examples

use sea_query::{*, tests_cfg::*};

let query = Query::select()
    .columns([Char::Character, Char::SizeW, Char::SizeH])
    .from(Char::Table)
    .and_where((Char::Table, Char::Character).into_column_ref().like("Ours'%"))
    .to_owned();

assert_eq!(
    query.to_string(MsSqlQueryBuilder),
    r#"SELECT [character], [size_w], [size_h] FROM [character] WHERE [character].[character] LIKE 'Ours''%'"#
);
assert_eq!(
    query.to_string(MysqlQueryBuilder),
    r#"SELECT `character`, `size_w`, `size_h` FROM `character` WHERE `character`.`character` LIKE 'Ours\'%'"#
);
assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE "character"."character" LIKE E'Ours\'%'"#
);
assert_eq!(
    query.to_string(SqliteQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE "character"."character" LIKE 'Ours''%'"#
);

Like with ESCAPE

use sea_query::{*, tests_cfg::*};

let query = Query::select()
    .columns([Char::Character, Char::SizeW, Char::SizeH])
    .from(Char::Table)
    .and_where((Char::Table, Char::Character).into_column_ref().like(LikeExpr::new(r"|_Our|_").escape('|')))
    .to_owned();

assert_eq!(
    query.to_string(MsSqlQueryBuilder),
    r#"SELECT [character], [size_w], [size_h] FROM [character] WHERE [character].[character] LIKE '|_Our|_' ESCAPE '|'"#
);
assert_eq!(
    query.to_string(MysqlQueryBuilder),
    r#"SELECT `character`, `size_w`, `size_h` FROM `character` WHERE `character`.`character` LIKE '|_Our|_' ESCAPE '|'"#
);
assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE "character"."character" LIKE '|_Our|_' ESCAPE '|'"#
);
assert_eq!(
    query.to_string(SqliteQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE "character"."character" LIKE '|_Our|_' ESCAPE '|'"#
);

fn lt<R>(self, right: R) -> Expr
where R: Into<Expr>,

Express a less than (<) expression.

Examples

use sea_query::{tests_cfg::*, *};

let query = Query::select()
    .columns([Char::Character, Char::SizeW, Char::SizeH])
    .from(Char::Table)
    .and_where((Char::Table, Char::SizeW).into_column_ref().lt(2))
    .to_owned();

assert_eq!(
    query.to_string(MsSqlQueryBuilder),
    r#"SELECT [character], [size_w], [size_h] FROM [character] WHERE [character].[size_w] < 2"#
);
assert_eq!(
    query.to_string(MysqlQueryBuilder),
    r#"SELECT `character`, `size_w`, `size_h` FROM `character` WHERE `character`.`size_w` < 2"#
);
assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE "character"."size_w" < 2"#
);
assert_eq!(
    query.to_string(SqliteQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE "character"."size_w" < 2"#
);

fn lte<R>(self, right: R) -> Expr
where R: Into<Expr>,

Express a less than or equal (<=) expression.

Examples

use sea_query::{tests_cfg::*, *};

let query = Query::select()
    .columns([Char::Character, Char::SizeW, Char::SizeH])
    .from(Char::Table)
    .and_where((Char::Table, Char::SizeW).into_column_ref().lte(2))
    .to_owned();

assert_eq!(
    query.to_string(MsSqlQueryBuilder),
    r#"SELECT [character], [size_w], [size_h] FROM [character] WHERE [character].[size_w] <= 2"#
);
assert_eq!(
    query.to_string(MysqlQueryBuilder),
    r#"SELECT `character`, `size_w`, `size_h` FROM `character` WHERE `character`.`size_w` <= 2"#
);
assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE "character"."size_w" <= 2"#
);
assert_eq!(
    query.to_string(SqliteQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE "character"."size_w" <= 2"#
);

fn modulo<R>(self, right: R) -> Expr
where R: Into<Expr>,

Express an arithmetic modulo operation.

Examples

use sea_query::{tests_cfg::*, *};

let query = Query::select()
    .columns([Char::Character, Char::SizeW, Char::SizeH])
    .from(Char::Table)
    .and_where(1.modulo(1).eq(2))
    .to_owned();

assert_eq!(
    query.to_string(MsSqlQueryBuilder),
    r#"SELECT [character], [size_w], [size_h] FROM [character] WHERE 1 % 1 = 2"#
);
assert_eq!(
    query.to_string(MysqlQueryBuilder),
    r#"SELECT `character`, `size_w`, `size_h` FROM `character` WHERE 1 % 1 = 2"#
);
assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE 1 % 1 = 2"#
);
assert_eq!(
    query.to_string(SqliteQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE 1 % 1 = 2"#
);

fn mul<R>(self, right: R) -> Expr
where R: Into<Expr>,

Express an arithmetic multiplication operation.

Examples

use sea_query::{tests_cfg::*, *};

let query = Query::select()
    .columns([Char::Character, Char::SizeW, Char::SizeH])
    .from(Char::Table)
    .and_where(1.mul(1).eq(2))
    .to_owned();

assert_eq!(
    query.to_string(MsSqlQueryBuilder),
    r#"SELECT [character], [size_w], [size_h] FROM [character] WHERE 1 * 1 = 2"#
);
assert_eq!(
    query.to_string(MysqlQueryBuilder),
    r#"SELECT `character`, `size_w`, `size_h` FROM `character` WHERE 1 * 1 = 2"#
);
assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE 1 * 1 = 2"#
);
assert_eq!(
    query.to_string(SqliteQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE 1 * 1 = 2"#
);

fn ne<R>(self, right: R) -> Expr
where R: Into<Expr>,

Express a not equal (<>) expression.

Examples

use sea_query::{*, tests_cfg::*};

let query = Query::select()
    .columns([Char::Character, Char::SizeW, Char::SizeH])
    .from(Char::Table)
    // Sometimes, you'll have to qualify the call because of conflicting std traits.
    .and_where(ExprTrait::ne("Morning", "Good"))
    .and_where(Char::Id.into_column_ref().ne(1))
    .to_owned();

assert_eq!(
    query.to_string(MsSqlQueryBuilder),
    r#"SELECT [character], [size_w], [size_h] FROM [character] WHERE 'Morning' <> 'Good' AND [id] <> 1"#
);
assert_eq!(
    query.to_string(MysqlQueryBuilder),
    r#"SELECT `character`, `size_w`, `size_h` FROM `character` WHERE 'Morning' <> 'Good' AND `id` <> 1"#
);
assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE 'Morning' <> 'Good' AND "id" <> 1"#
);
assert_eq!(
    query.to_string(SqliteQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE 'Morning' <> 'Good' AND "id" <> 1"#
);

fn not(self) -> Expr

Negates an expression with NOT.

Examples

use sea_query::{*, tests_cfg::*};

let query = Query::select()
    .columns([Char::Character, Char::SizeW, Char::SizeH])
    .from(Char::Table)
    .and_where(Expr::col((Char::Table, Char::SizeW)).is_null().not())
    .to_owned();

assert_eq!(
    query.to_string(MsSqlQueryBuilder),
    r#"SELECT [character], [size_w], [size_h] FROM [character] WHERE NOT [character].[size_w] IS NULL"#
);
assert_eq!(
    query.to_string(MysqlQueryBuilder),
    r#"SELECT `character`, `size_w`, `size_h` FROM `character` WHERE NOT `character`.`size_w` IS NULL"#
);
assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE NOT "character"."size_w" IS NULL"#
);
assert_eq!(
    query.to_string(SqliteQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE NOT "character"."size_w" IS NULL"#
);

fn not_between<A, B>(self, a: A, b: B) -> Expr
where A: Into<Expr>, B: Into<Expr>,

Express a NOT BETWEEN expression.

Examples

use sea_query::{*, tests_cfg::*};

let query = Query::select()
    .columns([Char::Character, Char::SizeW, Char::SizeH])
    .from(Char::Table)
    .and_where((Char::Table, Char::SizeW).into_column_ref().not_between(1, 10))
    .to_owned();

assert_eq!(
    query.to_string(MsSqlQueryBuilder),
    r#"SELECT [character], [size_w], [size_h] FROM [character] WHERE [character].[size_w] NOT BETWEEN 1 AND 10"#
);
assert_eq!(
    query.to_string(MysqlQueryBuilder),
    r#"SELECT `character`, `size_w`, `size_h` FROM `character` WHERE `character`.`size_w` NOT BETWEEN 1 AND 10"#
);
assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE "character"."size_w" NOT BETWEEN 1 AND 10"#
);
assert_eq!(
    query.to_string(SqliteQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE "character"."size_w" NOT BETWEEN 1 AND 10"#
);

fn not_equals<C>(self, col: C) -> Expr
where C: IntoColumnRef,

Express a not equal expression between two table columns, you will mainly use this to relate identical value between two table columns.

Examples

use sea_query::{*, tests_cfg::*};

let query = Query::select()
    .columns([Char::Character, Char::SizeW, Char::SizeH])
    .from(Char::Table)
    .and_where((Char::Table, Char::FontId).into_column_ref().not_equals((Font::Table, Font::Id)))
    .to_owned();

assert_eq!(
    query.to_string(MsSqlQueryBuilder),
    r#"SELECT [character], [size_w], [size_h] FROM [character] WHERE [character].[font_id] <> [font].[id]"#
);
assert_eq!(
    query.to_string(MysqlQueryBuilder),
    r#"SELECT `character`, `size_w`, `size_h` FROM `character` WHERE `character`.`font_id` <> `font`.`id`"#
);
assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE "character"."font_id" <> "font"."id""#
);
assert_eq!(
    query.to_string(SqliteQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE "character"."font_id" <> "font"."id""#
);

fn not_in_subquery(self, sel: SelectStatement) -> Expr

Express a NOT IN sub-query expression.

Examples

use sea_query::{*, tests_cfg::*};

let query = Query::select()
    .columns([Char::Character, Char::SizeW, Char::SizeH])
    .from(Char::Table)
    .and_where(Char::SizeW.into_column_ref().not_in_subquery(
        Query::select()
            .expr(Expr::cust("3 + 2 * 2"))
            .take()
    ))
    .to_owned();

assert_eq!(
    query.to_string(MsSqlQueryBuilder),
    r#"SELECT [character], [size_w], [size_h] FROM [character] WHERE [size_w] NOT IN (SELECT 3 + 2 * 2)"#
);
assert_eq!(
    query.to_string(MysqlQueryBuilder),
    r#"SELECT `character`, `size_w`, `size_h` FROM `character` WHERE `size_w` NOT IN (SELECT 3 + 2 * 2)"#
);
assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE "size_w" NOT IN (SELECT 3 + 2 * 2)"#
);
assert_eq!(
    query.to_string(SqliteQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE "size_w" NOT IN (SELECT 3 + 2 * 2)"#
);

fn not_like<L>(self, like: L) -> Expr
where L: IntoLikeExpr,

Express a NOT LIKE expression.

Examples

use sea_query::{*, tests_cfg::*};

let query = Query::select()
    .columns([Char::Character, Char::SizeW, Char::SizeH])
    .from(Char::Table)
    .and_where((Char::Table, Char::Character).into_column_ref().not_like("Ours'%"))
    .to_owned();

assert_eq!(
    query.to_string(MsSqlQueryBuilder),
    r#"SELECT [character], [size_w], [size_h] FROM [character] WHERE [character].[character] NOT LIKE 'Ours''%'"#
);
assert_eq!(
    query.to_string(MysqlQueryBuilder),
    r#"SELECT `character`, `size_w`, `size_h` FROM `character` WHERE `character`.`character` NOT LIKE 'Ours\'%'"#
);
assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE "character"."character" NOT LIKE E'Ours\'%'"#
);
assert_eq!(
    query.to_string(SqliteQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE "character"."character" NOT LIKE 'Ours''%'"#
);

fn or<R>(self, right: R) -> Expr
where R: Into<Expr>,

Express a logical OR operation.

Examples

use sea_query::{tests_cfg::*, *};

let query = Query::select()
    .columns([Char::Character, Char::SizeW, Char::SizeH])
    .from(Char::Table)
    .and_where(false.or(true))
    .to_owned();

assert_eq!(
    query.to_string(MysqlQueryBuilder),
    r#"SELECT `character`, `size_w`, `size_h` FROM `character` WHERE FALSE OR TRUE"#
);
assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE FALSE OR TRUE"#
);
assert_eq!(
    query.to_string(SqliteQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE FALSE OR TRUE"#
);

fn right_shift<R>(self, right: R) -> Expr
where R: Into<Expr>,

Express a bitwise right shift.

Examples

use sea_query::{tests_cfg::*, *};

let query = Query::select()
    .columns([Char::Character, Char::SizeW, Char::SizeH])
    .from(Char::Table)
    .and_where(1.right_shift(1).eq(2))
    .to_owned();

assert_eq!(
    query.to_string(MsSqlQueryBuilder),
    r#"SELECT [character], [size_w], [size_h] FROM [character] WHERE 1 >> 1 = 2"#
);
assert_eq!(
    query.to_string(MysqlQueryBuilder),
    r#"SELECT `character`, `size_w`, `size_h` FROM `character` WHERE 1 >> 1 = 2"#
);
assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE 1 >> 1 = 2"#
);
assert_eq!(
    query.to_string(SqliteQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE 1 >> 1 = 2"#
);

fn sub<R>(self, right: R) -> Expr
where R: Into<Expr>,

Express an arithmetic subtraction operation.

Examples

use sea_query::{tests_cfg::*, *};

let query = Query::select()
    .columns([Char::Character, Char::SizeW, Char::SizeH])
    .from(Char::Table)
    .and_where(1.sub(1).eq(2))
    .to_owned();

assert_eq!(
    query.to_string(MsSqlQueryBuilder),
    r#"SELECT [character], [size_w], [size_h] FROM [character] WHERE 1 - 1 = 2"#
);
assert_eq!(
    query.to_string(MysqlQueryBuilder),
    r#"SELECT `character`, `size_w`, `size_h` FROM `character` WHERE 1 - 1 = 2"#
);
assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE 1 - 1 = 2"#
);
assert_eq!(
    query.to_string(SqliteQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE 1 - 1 = 2"#
);

fn bit_and<R>(self, right: R) -> Expr
where R: Into<Expr>,

Express a bitwise AND operation.

Examples

use sea_query::{tests_cfg::*, *};

let query = Query::select().expr(1.bit_and(2).eq(3)).to_owned();

assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT (1 & 2) = 3"#
);

let query = Query::select()
    .columns([Char::Character, Char::SizeW, Char::SizeH])
    .from(Char::Table)
    .and_where(1.bit_and(1).eq(1))
    .to_owned();

assert_eq!(
    query.to_string(MysqlQueryBuilder),
    r#"SELECT `character`, `size_w`, `size_h` FROM `character` WHERE (1 & 1) = 1"#
);
assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE (1 & 1) = 1"#
);
assert_eq!(
    query.to_string(SqliteQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE (1 & 1) = 1"#
);

fn bit_or<R>(self, right: R) -> Expr
where R: Into<Expr>,

Express a bitwise OR operation.

Examples

use sea_query::{tests_cfg::*, *};

let query = Query::select()
    .columns([Char::Character, Char::SizeW, Char::SizeH])
    .from(Char::Table)
    .and_where(1.bit_or(1).eq(1))
    .to_owned();

assert_eq!(
    query.to_string(MysqlQueryBuilder),
    r#"SELECT `character`, `size_w`, `size_h` FROM `character` WHERE (1 | 1) = 1"#
);
assert_eq!(
    query.to_string(PostgresQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE (1 | 1) = 1"#
);
assert_eq!(
    query.to_string(SqliteQueryBuilder),
    r#"SELECT "character", "size_w", "size_h" FROM "character" WHERE (1 | 1) = 1"#
);

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.