sea_query::query

Module select

Structs

JoinExpr

Join expression used in select statement

LockClause

SelectExpr

Select expression used in select statement

SelectInto

The INTO clause in SELECT statement

SelectStatement

Select rows from an existing table

Enums

LockBehavior

List of lock behavior can be used in select statement

LockType

List of lock types that can be used in select statement

SelectDistinct

List of distinct keywords that can be used in select statement

SelectIntoTableModifier

The table type of target table of SELECT INTO clause

UnionType

List of union types that can be used in union clause

WindowSelectType

Window type in SelectExpr

Traits

SelectExprTrait

Extension methods for building a SelectExpr from an expression.