sea_query::query

Module with

Structs

CommonTableExpression

A table definition inside a WITH clause (WithClause).

Cycle

For recursive WithQuery WithClauses the CYCLE sql clause can be specified to avoid creating an infinite traversals that loops on graph cycles indefinitely.

Search

For recursive WithQuery WithClauses the traversing order can be specified in some databases that support this functionality.

WithClause

A WITH clause can contain one or multiple common table expressions (CommonTableExpression).

WithQuery

A WITH query. A simple SQL query that has a WITH clause (WithClause).

Enums

CteQuery 🔒

SearchOrder

For recursive WithQuery WithClauses the traversing order can be specified in some databases that support this functionality.