Trait sea_query::schema::SchemaStatementBuilder

pub trait SchemaStatementBuilder {
    fn build<T: SchemaBuilder>(&self, schema_builder: T) -> String;
    fn build_any(&self, schema_builder: &dyn SchemaBuilder) -> String;

    fn to_string<T: SchemaBuilder>(&self, schema_builder: T) -> String { ... }
}

Required Methods

fn build<T: SchemaBuilder>(&self, schema_builder: T) -> String

Build corresponding SQL statement for certain database backend and return SQL string

fn build_any(&self, schema_builder: &dyn SchemaBuilder) -> String

Build corresponding SQL statement for certain database backend and return SQL string

Provided Methods

fn to_string<T: SchemaBuilder>(&self, schema_builder: T) -> String

Build corresponding SQL statement for certain database backend and return SQL string