sea_query

Module types

Base types used throughout sea-query.

Structs

Alias

An explicit wrapper for Idens which are dynamic user-provided strings.

Asterisk

Asterisk (“*”)

ColumnName

A column name, potentially qualified as (database.)(schema.)(table.)column.

DatabaseName

An identifier that represents a database name.

DynIden

A prepared (quoted) identifier string.

LikeExpr

Like Expression

NullAlias

Null Alias

OrderExpr

Order expression

Quote

SchemaName

A schema name, potentially qualified as (database.)schema.

SeaRc

A legacy namespace for compatibility.

TableName

A table name, potentially qualified as (database.)(schema.)table.

TypeRef

An SQL type name, potentially qualified as (database.)(schema.)type.

Enums

BinOper

Binary operators.

ColumnRef

Column references.

JoinOn

Join on types

JoinType

Join types

Keyword

Known SQL keywords that can be used as expressions.

LogicalChainOper

Logical chain operator: conjunction or disjunction.

NullOrdering

Nulls order

Order

Ordering options

SubQueryOper

SubQuery operators

TableRef

Table references

UnOper

Unary operators.

Traits

Iden

Identifier

IdenList

IdenStatic

Identifier statically known at compile-time.

IntoColumnRef

IntoIden

IntoLikeExpr

IntoTableRef

IntoTypeRef

MaybeQualifiedOnce

A name that can be unqualified (foo) or qualified once (foo.bar).

MaybeQualifiedTwice

A name that can be unqualified (foo), qualified once (foo.bar), or twice (foo.bar.baz).

Functions

is_static_iden

Return whether this identifier needs to be escaped. Right now we’re very safe and only return true for identifiers composed of a-zA-Z0-9_.

Type Aliases

RcOrArc

A reference counted pointer: either Rc or Arc, depending on the feature flags.