sea_query::types::iden

Module core

“Core”, low-level identifier types.

Structs

Alias

An explicit wrapper for Idens which are dynamic user-provided strings.

Asterisk

Asterisk (“*”)

DynIden

A prepared (quoted) identifier string.

NullAlias

Null Alias

Traits

Iden

Identifier

IdenStatic

Identifier statically known at compile-time.

IntoIden

Functions

is_static_iden

Return whether this identifier needs to be escaped. Right now we’re very safe and only return true for identifiers composed of a-zA-Z0-9_.