sea_query::schema

Trait SchemaStatementBuilder

pub trait SchemaStatementBuilder {
    // Required methods
    fn build<T: SchemaBuilder>(&self, schema_builder: T) -> String;
    fn build_any(&self, schema_builder: &dyn SchemaBuilder) -> String;

    // Provided method
    fn to_string<T: SchemaBuilder>(&self, schema_builder: T) -> String { ... }
}

Required Methods

fn build<T: SchemaBuilder>(&self, schema_builder: T) -> String

Build corresponding SQL statement for certain database backend and return SQL string

fn build_any(&self, schema_builder: &dyn SchemaBuilder) -> String

Build corresponding SQL statement for certain database backend and return SQL string

Provided Methods

fn to_string<T: SchemaBuilder>(&self, schema_builder: T) -> String

Build corresponding SQL statement for certain database backend and return SQL string

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.