sea_query::schema

Trait SchemaStatementBuilder

pub trait SchemaStatementBuilder {
    // Required method
    fn build<T>(&self, schema_builder: T) -> String
       where T: SchemaBuilder;

    // Provided method
    fn to_string<T>(&self, schema_builder: T) -> String
       where T: SchemaBuilder { ... }
}

Required Methods

fn build<T>(&self, schema_builder: T) -> String
where T: SchemaBuilder,

Build corresponding SQL statement for certain database backend and return SQL string

Provided Methods

fn to_string<T>(&self, schema_builder: T) -> String
where T: SchemaBuilder,

Build corresponding SQL statement for certain database backend and return SQL string

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.